s3_lambda_ecs/requirements.txt
attrs==19.1.0
aws-cdk.assets==1.8.0
aws-cdk.aws-apigateway==1.8.0
aws-cdk.aws-applicationautoscaling==1.8.0
aws-cdk.aws-autoscaling==1.7.0
aws-cdk.aws-autoscaling-common==1.8.0
aws-cdk.aws-autoscaling-hooktargets==1.7.0
aws-cdk.aws-certificatemanager==1.8.0
aws-cdk.aws-cloudformation==1.8.0
aws-cdk.aws-cloudfront==1.7.0
aws-cdk.aws-cloudwatch==1.8.0
aws-cdk.aws-dynamodb==1.8.0
aws-cdk.aws-ec2==1.8.0
aws-cdk.aws-ecr==1.7.0
aws-cdk.aws-ecr-assets==1.7.0
aws-cdk.aws-ecs==1.7.0
aws-cdk.aws-elasticloadbalancing==1.7.0
aws-cdk.aws-elasticloadbalancingv2==1.8.0
aws-cdk.aws-events==1.8.0
aws-cdk.aws-iam==1.8.0
aws-cdk.aws-kinesis==1.8.0
aws-cdk.aws-kms==1.8.0
aws-cdk.aws-lambda==1.8.0
aws-cdk.aws-lambda-event-sources==1.8.0
aws-cdk.aws-logs==1.8.0
aws-cdk.aws-route53==1.8.0
aws-cdk.aws-route53-targets==1.7.0
aws-cdk.aws-s3==1.8.0
aws-cdk.aws-s3-assets==1.8.0
aws-cdk.aws-s3-notifications==1.8.0
aws-cdk.aws-secretsmanager==1.7.0
aws-cdk.aws-servicediscovery==1.7.0
aws-cdk.aws-sns==1.8.0
aws-cdk.aws-sns-subscriptions==1.8.0
aws-cdk.aws-sqs==1.8.0
aws-cdk.aws-ssm==1.8.0
aws-cdk.core==1.8.0
aws-cdk.cx-api==1.8.0
aws-cdk.region-info==1.8.0
cattrs==0.9.0
jsii==0.16.0
publication==0.0.3
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
